fix codestyle

+++ main.js
@@ -529,11 +529,9 @@ if (fs.existsSync(haDisableFile)) app.disableHardwareAcceleration();
 const cduFolderPath = path.join(app.getPath('userData'), 'CrashMemoryDump');
 const cduDisableFile = path.join(app.getPath('userData'), 'CDUploadDisable');
 if (fs.existsSync(cduDisableFile)) {
-  if (fs.existsSync(cduFolderPath))
-    rimraf.sync(cduFolderPath);
+  if (fs.existsSync(cduFolderPath)) rimraf.sync(cduFolderPath);
 } else {
-  if (!fs.existsSync(cduFolderPath))
-    fs.mkdirSync(cduFolderPath);
+  if (!fs.existsSync(cduFolderPath)) fs.mkdirSync(cduFolderPath);
 }
 
 app.setAsDefaultProtocolClient('slobs');
